+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - If waiting for Ceph health to become "OK" times out during the ``platform-services``
   stage of ``sat bootsys boot``, the Ceph services will now be restarted on the 
   storage nodes, and Ceph health will be waited on again.
+- Fixed an error in the ``platform-services`` stage of ``sat bootsys boot``
+  related to trying to start non-existent Ceph services.
 
 ## [3.7.1] - 2021-07-27